requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
streamlit==1.28.0
pydantic==2.4.2
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
gunicorn==21.2.0
httpx==0.25.1
